WIP moving code from BaseMainActivity to MineViewGL
